核销bug 修复

+++ app/Api/Controllers/Bill/ChargeController.php
@@ -534,7 +534,10 @@ class ChargeController extends BaseController
 		$request->validate([
 			'id' => 'required|numeric|gt:0',
 		], $msg);
-
+		$exists = $this->chargeService->chargeRecord()->where('id', $request->id)->exists();
+		if ($exists) {
+			return $this->error("核销记录不存在或者已删除");
+		}
 		$res = $this->chargeService->deleteChargeRecord(($request->id));
 		return $res ? $this->success("核销删除成功") : $this->error("删除失败");
 	}
